+++ src/bloqade/ir/control/traits/hash.py
